+++ src/configs/import.ts
@@ -6,9 +6,6 @@
 
 module.exports = {
   rules: {
-    'comma-dangle': 'off', // NOTE: this is handled by @typescript-eslint/comma-dangle
-    'no-dupe-keys': 'error',
-    'no-void': ['error', { allowAsStatement: true }],
     'import/order': ['error', {
       'newlines-between': 'always',
       groups: [