+++ src/configs/typescript.ts
@@ -10,7 +10,6 @@ module.exports = {
       default: 'array',
       readonly: 'array',
     }],
-    semi: 'off',
     '@typescript-eslint/comma-dangle': ['error', {
       arrays: 'always-multiline',
       objects: 'always-multiline',
@@ -21,8 +20,6 @@ module.exports = {
       generics: 'ignore',
       tuples: 'always-multiline',
     }],
-    '@typescript-eslint/semi': ['error', 'never'],
-    'no-dupe-keys': 'error',
     '@typescript-eslint/keyword-spacing': ['error'],
     '@typescript-eslint/naming-convention': [
       'error',
@@ -50,6 +47,7 @@ module.exports = {
         ],
       },
     ],
+    '@typescript-eslint/no-explicit-any': ['error'],
     '@typescript-eslint/member-delimiter-style': ['error', {
       multiline: {
         delimiter: 'comma',
@@ -60,23 +58,11 @@ module.exports = {
         requireLast: false,
       },
     }],
-    'no-return-await': 'off',
-    '@typescript-eslint/return-await': [
-      'error',
-      'in-try-catch',
-    ],
-    '@typescript-eslint/strict-boolean-expressions': ['off'],
     '@typescript-eslint/consistent-type-definitions': 'off',
-    '@typescript-eslint/consistent-type-imports': ['error'],
-    '@typescript-eslint/ban-ts-comment': [
-      'error', {
-        'ts-expect-error': 'allow-with-description',
-        minimumDescriptionLength: 10,
-      },
-    ],
     '@typescript-eslint/ban-types': [
       'error',
       {
+        extendDefaults: true,
         types: {
           object: {
             message: [
@@ -84,22 +70,6 @@ module.exports = {
               'Consider using `Record<string, unknown>` instead, as it allows you to more easily inspect and use the keys.',
             ].join('\n'),
             fixWith: 'Record<string, unknown>',
-          },
-          Object: {
-            message: [
-              'The `Object` type actually means "any non-nullish value", so it is marginally better than `unknown`.',
-              '- If you want a type meaning "any object", you probably want `Record<string, unknown>` instead.',
-              '- If you want a type meaning "any value", you probably want `unknown` instead.',
-            ].join('\n'),
-            fixWith: 'Record<string, unknown>',
-          },
-          Function: {
-            message: [
-              'The `Function` type accepts any function-like value.',
-              'It provides no type safety when calling the function, which can be a common source of bugs.',
-              'It also accepts things like class declarations, which will throw at runtime as they will not be called with `new`.',
-              'If you are expecting the function to accept certain arguments, you should explicitly define the function shape.',
-            ].join('\n'),
           },
         },
       },